refactor(utils): give the range-overlap predicate one home
The core, the agenda plugin and the recurrence plugin each carried their own
copy of the same three-case overlap test. That is what `architecture.md`
forbids, and what these copies did in practice: the grid showed a one-day event
while the agenda showed two, and the recurrence filter went wrong twice in two
days, once in each direction.

`overlapsRange` moves to `@ilamy/utils/helpers` with its tests, and all three
call it. It is typed over `{ start: Dayjs; end: Dayjs }` rather than
`CalendarEvent`, so utils needs no dependency on `@ilamy/types` and cells,
selections and view ranges fit the same helper; `CalendarEvent` satisfies it
structurally, so no call site changed shape. The name drops its `event` prefix
for the same reason.

The agenda keeps its own rule about which days an event appears under (all-day
repeats per covered day, timed appears once under its start day) and delegates
only the overlap half.

Refs #248
