+++ docs/logs/2026-08-15.md
@@ -46,3 +46,28 @@ Two more findings, both verified by measurement first.
 **Both findings were the same shape**: a strict end comparison that forgot a zero-duration event has no end to speak of. After fixing them, the sweep for that shape came back clean: `pipeline` already keeps such an event through `startsInRange`, `horizontal` classifies it single-unit and paints it, `dnd-utils` makes no comparison.
 
 **The three-case overlap predicate now exists in three places** (`lib/events/pipeline.ts`, the agenda's `appearsOnDay`, the recurrence range filter). That is precisely the duplication `.agents/rules/change-impact.md` names as the cause of this whole class of bug, and `.agents/rules/architecture.md` says shared logic belongs in `@ilamy/utils`. Relocating `eventOverlapsRange` there, with both plugins importing it, is the follow-up that would stop these three drifting apart again. Not done here: it touches three packages and is a refactor rather than a review fix.
+
+## Follow-up: one home for the overlap predicate
+
+- **[utils]**: `overlapsRange` now lives in `@ilamy/utils/helpers`, and the core, the agenda plugin and the recurrence plugin all call it. It had grown three private copies, which is what `.agents/rules/architecture.md` forbids and what `.agents/rules/change-impact.md` names as the cause of this whole bug class: the grid said a one-day event, the agenda said two.
+
+- **[utils]**: Typed over `{ start: Dayjs; end: Dayjs }` rather than `CalendarEvent`, so `@ilamy/utils` needs no dependency on `@ilamy/types`. `CalendarEvent` satisfies it structurally, so no call site changed shape. The name lost its `event` prefix because the helper knows nothing about events: cells, selections and view ranges are the same shape.
+
+- **[agenda]**: `appearsOnDay` keeps its own rule (all-day repeats under each covered day, timed appears once under its start day) and delegates only the overlap half.
+
+## Files Modified (follow-up)
+
+- `packages/utils/src/helpers.ts` - `overlapsRange` with the boundary contract documented
+- `packages/utils/src/helpers.test.ts` - NEW; the eight cases moved from the calendar package with the code
+- `packages/calendar/src/lib/events/pipeline.ts` / `.test.ts` - predicate and its tests removed; resource helpers stay
+- `packages/calendar/src/features/plugins/lib/create-plugin-runtime.ts`, `hooks/useProcessedWeekEvents.ts` - import the shared helper
+- `packages/plugins/agenda/src/utils/group-events-by-day.ts`, `packages/plugins/recurrence/src/utils/generate-recurring-events.ts` - inline copies deleted
+- `AGENTS.md`, `docs/types-and-interfaces.md` - the predicate's new home
+
+## Notes (follow-up)
+
+**Test bookkeeping**: the calendar package went 910 to 902 and utils 9 to 17, so all eight cases moved and none were duplicated or lost. Counted rather than assumed, since a silently dropped test is exactly what a move like this hides.
+
+**A new test file was created** (`packages/utils/src/helpers.test.ts`), which the TDD rule normally forbids. Leaving the tests in the calendar package to exercise code that now lives in utils would be worse: the tests move with the code.
+
+**Stale references swept**: comments in both plugin test files, `docs/types-and-interfaces.md`, and the two `AGENTS.md` Key Paths lines. Historical documents (dev logs, the v2 plan docs, `agenda-plugin-plan.md`) still name `lib/events/pipeline.ts` and are left alone: they record what was true when written.
